Refactor UnsupervisedLoss into its own module.
Moved the UnsupervisedLoss function from base_metrics.py to a new unsupervised_loss.py module for better separation of concerns and modularity. Updated imports and references across related files to align with the new structure.
